% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom4R.R
\docType{package}
\name{atom4R}
\alias{atom4R}
\alias{atom4R-package}
\title{Tools to Handle and Publish Metadata as Atom XML Format}
\description{
Provides tools to read/write/publish metadata based on the Atom XML syndication format.
This includes support of Dublin Core XML implementation, and a client to APIs implementing the AtomPub
SWORD API specification.
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
