% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomAbstractObject.R
\docType{class}
\name{AtomAbstractObject}
\alias{AtomAbstractObject}
\title{Atom feed class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Atom abstract Object
}
\description{
This class models an atom abstract object
}
\details{
AtomAbstractObject
}
\note{
abstract class used internally by \pkg{atom4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to create an Atom abstract object
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{atom}
