% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomPerson.R
\docType{class}
\name{AtomAuthor}
\alias{AtomAuthor}
\title{Atom Author class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Atom Author
}
\description{
This class models an Atom Author
}
\details{
AtomAuthor
}
\section{Fields}{

\describe{
\item{\code{name}}{Person name}

\item{\code{uri}}{Person uri}

\item{\code{email}}{Person email}
}}

\section{Methods inherited from \code{AtomPerson}}{

\describe{
 \item{\code{new(xml, name, uri, email)}}{
   This method is used to create an Atom
 }
 \item{\code{setName(name)}}{
   Set person name
 }
 \item{\code{setUri(uri)}}{
   Set person uri
 }
 \item{\code{setEmail(email)}}{
   Set person email
 }
}
}

\examples{
\dontrun{
  author <- AtomAuthor$new(name "John Doe", email = "john.doe@atom4R.com")
}

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Atom}
\keyword{Author}
