% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomCategory.R
\docType{class}
\name{AtomCategory}
\alias{AtomCategory}
\title{Atom Category class}
\format{
\code{\Category{R6Class}} object.
}
\value{
Object of \code{\Category{R6Class}} for modelling an Atom Category
}
\description{
This class models an atom Category
}
\details{
AtomCategory
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to create an Atom Category
 }
 \item{\code{setHref(href)}}{
   Sets the href
 }
 \item{\code{setHreflang(hreflang)}}{
   Sets the href lang
 }
 \item{\code{setRel(rel)}}{
   Sets the rel
 }
 \item{\code{setType(type)}}{
   Sets the type
 }
 \item{\code{setTitle(title)}}{
   Sets the title
 }
 \item{\code{setLength(length)}}{
   Sets the length
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Category}
\keyword{atom}
