% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomFeed.R
\docType{class}
\name{AtomFeed}
\alias{AtomFeed}
\title{Atom feed class}
\format{\code{\link{R6Class}} object.}
\usage{
AtomFeed
}
\value{
Object of \code{\link{R6Class}} for modelling an Atom feed
}
\description{
This class models an atom feed
}
\details{
AtomFeed
}
\section{Fields}{

\describe{
\item{\code{entries}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to create an Atom
 }
}
}

\examples{
\dontrun{
}

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{atom}
\keyword{feed}
