% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomLink.R
\docType{class}
\name{AtomLink}
\alias{AtomLink}
\title{Atom Link class}
\format{\code{\link{R6Class}} object.}
\usage{
AtomLink
}
\value{
Object of \code{\link{R6Class}} for modelling an Atom Link
}
\description{
This class models an atom Link
}
\details{
AtomLink
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to create an Atom Link
 }
 \item{\code{setHref(href)}}{
   Sets the href
 }
 \item{\code{setHreflang(hreflang)}}{
   Sets the href lang
 }
 \item{\code{setRel(rel)}}{
   Sets the rel
 }
 \item{\code{setType(type)}}{
   Sets the type
 }
 \item{\code{setTitle(title)}}{
   Sets the title
 }
 \item{\code{setLength(length)}}{
   Sets the length
 }
}
}

\examples{
\dontrun{
}

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Link}
\keyword{atom}
