% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomNamespace.R
\docType{class}
\name{AtomNamespace}
\alias{AtomNamespace}
\title{AtomNamespace}
\format{\code{\link{R6Class}} object.}
\usage{
AtomNamespace
}
\value{
Object of \code{\link{R6Class}} for modelling an Atom Namespace
}
\description{
AtomNamespace
}
\note{
ISO class used internally by atom4R for specifying XML namespaces
}
\section{Methods}{

\describe{
 \item{\code{new(id, uri)}}{
   This method is used to instantiate an Atom
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{metadata}
\keyword{namespace}
