% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomPubClient.R
\docType{class}
\name{AtomPubClient}
\alias{AtomPubClient}
\title{AtomPubClient class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an AtomPub client
}
\description{
This class models an AtomPub service client
}
\details{
AtomPubClient
}
\note{
Abstract class used internally for AtomPub (Atom Publishing Protocol) clients
}
\section{Methods}{

\describe{
 \item{\code{new(url, user, password, token)}}{
   This method is to instantiate an AtomPub Client
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Atom}
\keyword{Person}
