% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomAbstractObject.R
\name{cacheAtomClasses}
\alias{cacheAtomClasses}
\title{cacheAtomClasses}
\usage{
cacheAtomClasses()
}
\description{
\code{\link{cacheAtomClasses}} allows to cache the list of
\pkg{atom4R} classes or extended. This is especially required to fasten
the decoding of metadata elements from an XML file. It is called internally
by \pkg{atom4R} the first function \code{\link{getAtomClasses}}.
}
\examples{
  cacheAtomClasses()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
