% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCAbstract}
\alias{DCAbstract}
\title{DCAbstract}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core 'abstract' element
}
\description{
This class models an DublinCore 'abstract' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, value)}}{
   This method is used to create an Dublin core 'abstract' element
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/abstract
}
\keyword{'abstract'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
