% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCAccrualPolicy}
\alias{DCAccrualPolicy}
\title{DCAccrualPolicy}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core 'accrualPolicy' element
}
\description{
This class models an DublinCore 'accrualPolicy' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, value)}}{
   This method is used to create an Dublin core 'accrualPolicy' element
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/accrualPolicy
}
\keyword{'accrualPolicy'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
