% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCAvailable}
\alias{DCAvailable}
\title{DCAvailable}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core 'available' element
}
\description{
This class models an DublinCore 'available' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, value)}}{
   This method is used to create an Dublin core 'available' element
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/available
}
\keyword{'available'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
