% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCBibliographicCitation}
\alias{DCBibliographicCitation}
\title{DCBibliographicCitation}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core 'bibliographicCitation' element
}
\description{
This class models an DublinCore 'bibliographicCitation' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, value)}}{
   This method is used to create an Dublin core 'bibliographicCitation' element
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/bibliographicCitation
}
\keyword{'bibliographicCitation'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
