% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCContributor}
\alias{DCContributor}
\title{DCContributor}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core 'contributor' element
}
\description{
This class models an DublinCore 'contributor' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, value, dc)}}{
   This method is used to create an Dublin core 'contributor' element.
   Use \code{dc} to \code{TRUE} to use Dublin core namespace instead of DC terms.
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/contributor
}
\keyword{'contributor'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
