% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCCreated}
\alias{DCCreated}
\title{DCCreated}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core Terms 'date' element
}
\description{
This class models an DublinCore Terms 'date' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, value)}}{
   This method is used to create an Dublin core Terms 'date' element
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/created
}
\keyword{'date'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
