% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCDate}
\alias{DCDate}
\title{DCDate}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core 'date' element
}
\description{
This class models an DublinCore 'date' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, term, value, dc)}}{
   This method is used to create an Dublin core Terms 'date' element.
   Use \code{dc} to \code{TRUE} to use Dublin core namespace instead of DC terms.
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/date
}
\keyword{'date'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
