% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCDateAccepted}
\alias{DCDateAccepted}
\title{DCDateAccepted}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core 'dateAccepted' element
}
\description{
This class models an DublinCore 'dateAccepted' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, value)}}{
   This method is used to create an Dublin core Terms 'dateAccepted' element.
   Use \code{dc} to \code{TRUE} to use Dublin core namespace instead of DC terms.
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/dateAccepted
}
\keyword{'dateAccepted'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
