% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCDateCopyrighted}
\alias{DCDateCopyrighted}
\title{DCDateCopyrighted}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core 'dateCopyrighted' element
}
\description{
This class models an DublinCore 'dateCopyrighted' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, value)}}{
   This method is used to create an Dublin core Terms 'dateCopyrighted' element
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/dateCopyrighted
}
\keyword{'dateCopyrighted'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
