% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCDateSubmitted}
\alias{DCDateSubmitted}
\title{DCDateSubmitted}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core 'dateSubmitted' element
}
\description{
This class models an DublinCore 'dateSubmitted' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, value)}}{
   This method is used to create an Dublin core Terms 'dateSubmitted' element
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/dateSubmitted
}
\keyword{'dateSubmitted'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
