% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\docType{class}
\name{DCElement}
\alias{DCElement}
\title{DublinCore element class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core element
}
\description{
This class models an DublinCore element
}
\details{
DCElement
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\note{
Class used internally by \pkg{atom4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xml, term, value)}}{
   This method is used to create an Dublin core element
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
