% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCEntry.R
\docType{class}
\name{DCEntry}
\alias{DCEntry}
\title{Dublin Core Entry class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core Entry
}
\description{
This class models an Dublin Core Entry
}
\details{
DCEntry
}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to create an Dublin Core
 }
}
}

\examples{
\dontrun{
}

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Core}
\keyword{Dublin}
\keyword{Entry}
\keyword{dc}
