% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCExtent}
\alias{DCExtent}
\title{DCExtent}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core 'extent' element
}
\description{
This class models an DublinCore 'extent' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, value)}}{
   This method is used to create an Dublin core Terms 'extent' element
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/extent
}
\keyword{'extent'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
