% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCRightsHolder}
\alias{DCRightsHolder}
\title{DCRightsHolder}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core Terms 'rightsHolder' element
}
\description{
This class models an DublinCore 'rightsHolder' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, value)}}{
   This method is used to create an Dublin core Terms 'rightsHolder' element
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/rightsHolder
}
\keyword{'rightsHolder'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
