% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCElement.R
\name{DCTitle}
\alias{DCTitle}
\title{DCTitle}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core Terms 'title' element
}
\description{
This class models an DublinCore 'title' element
}
\section{Methods}{

\describe{
 \item{\code{new(xml, term, value, dc)}}{
   This method is used to create an Dublin core Terms 'title' element.
   Use \code{dc} to \code{TRUE} to use Dublin core namespace instead of DC terms.
 }
}
}

\references{
Dublin Core Metadata Initiative. https://www.dublincore.org/specifications/dublin-core/dcmi-terms/terms/title
}
\keyword{'title'}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
