% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomAbstractObject.R
\name{getAtomClasses}
\alias{getAtomClasses}
\title{getAtomClasses}
\usage{
getAtomClasses()
}
\description{
get the list of Atom classes, ie classes extending \link{AtomAbstractObject} super class,
including classes eventually defined outside \pkg{atom4R}. In case the latter is on the search path,
the list of Atom classes will be cached for optimized used by \pkg{atom4R} encoder/decoder.
}
\examples{
  getAtomClasses()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
