% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomNamespace.R
\name{getAtomNamespace}
\alias{getAtomNamespace}
\title{getAtomNamespace}
\usage{
getAtomNamespace(id)
}
\arguments{
\item{id}{namespace prefix}
}
\description{
\code{getAtomNamespace} gets a namespace given its id
}
\examples{
  getAtomNamespace("GMD")

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
