% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomNamespace.R
\name{getAtomNamespaces}
\alias{getAtomNamespaces}
\title{getAtomNamespaces}
\usage{
getAtomNamespaces()
}
\description{
\code{getAtomNamespaces} gets the list of namespaces registered
}
\examples{
  getAtomNamespaces()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
