% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomSchemas.R
\name{getAtomSchemas}
\alias{getAtomSchemas}
\title{getAtomSchemas}
\usage{
getAtomSchemas()
}
\description{
\code{getAtomSchemas} gets the schemas registered in \pkg{atom4R}
}
\examples{
  getAtomSchemas()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
