% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCMIVocabulary.R
\name{getDCMIVocabularies}
\alias{getDCMIVocabularies}
\title{getDCMIVocabularies}
\usage{
getDCMIVocabularies()
}
\description{
\code{getDCMIVocabularies} allows to get the list of DCMI Vocabularies
registered in \pkg{atom4R}
}
\examples{
  getDCMIVocabularies()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
