% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCMIVocabulary.R
\name{getDCMIVocabulary}
\alias{getDCMIVocabulary}
\title{getDCMIVocabulary}
\usage{
getDCMIVocabulary(id)
}
\arguments{
\item{id}{identifier of the vocabulary}
}
\description{
\code{getDCMIVocabulary} allows to get a registered DCMI Vocabulary by id
registered in \pkg{atom4R}
}
\examples{
  getDCMIVocabulary(id = "http://purl.org/dc/dcmitype/")

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
