% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomSchemas.R
\name{registerAtomSchema}
\alias{registerAtomSchema}
\title{registerAtomSchema}
\usage{
registerAtomSchema(xsdFile)
}
\arguments{
\item{xsdFile}{the schema XSD file}
}
\description{
\code{registerAtomSchema} allows to register a new schema
in \pkg{atom4R}
}
\examples{
  atom_xsd_file <- system.file("extdata/schemas/atom/atom.xsd", package = "atom4R")
  registerAtomSchema(xsdFile = atom_xsd_file)

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
