% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwordClient.R
\docType{class}
\name{SwordDataverseClient}
\alias{SwordDataverseClient}
\title{SWORD Dataverse client class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Sword Dataverse-specific APIclient
}
\description{
This class models an Sword service Dataverse-specific  API client
}
\details{
SwordDataverseClient
}
\section{Methods inherited from \code{SwordClient}}{

\describe{
 \item{\code{getServiceDocument()}}{
   Gets a representation in R of the SWORD service document (capabilities)
 }
}
}

\section{Methods}{

\describe{
 \item{\code{new(url, token, logger)}}{
   This method is to instantiate an Sword API Dataverse-specific Client.
   The \code{logger} allows to specify the level of log (default is NULL), either "INFO"
   for \pkg{atom4R} logs or "DEBUG" for verbose HTTP client (curl) logs.
 }
 \item{\code{getCollectionMembers(collectionId)}}{
   List the collection members, ie list of entries for a dataverse.
 }
 \item{\code{getDataverse(pretty)}}{
   List the dataverses, equivalent to \code{listCollections()}. The \code{pretty} argument
   can be set to \code{TRUE} to retrieve a \code{data.frame}, otherwise a \code{list} is
   returned.
 }
 \item{\code{getDataverse(dataverse)}}{
   Get a dataverse by ID
 }
 \item{\code{editDataverseEntry(identifier)}}{
   Edits a dataverse entry by identifier
 }
 \item{\code{getDataverseEntry(identifier)}}{
   Gets a dataverse entry by identifier
 }
 \item{\code{createDataverseEntry(dataverse, entry)}}{
   Creates a dataverse entry in the target \code{dataverse}. The entry should be an object
   of class \code{AtomEntry} or \code{DCEntry} (Dublin core entry).
 }
 \item{\code{updateDataverseEntry(dataverse, entry, doi)}}{
   Update a dataverse entry in the target \code{dataverse}. The entry should be an object
   of class \code{AtomEntry} or \code{DCEntry} (Dublin core entry). To update the entry it
   is necessary to specify the \code{doi} of the entry
 }
 \item{\code{deleteDataverseEntry(identifier)}}{
   Deletes a dataverse entry by identifier
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{API}
\keyword{Client}
\keyword{Dataverse}
\keyword{SWORD}
