% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom4R.R
\docType{package}
\name{atom4R}
\alias{atom4R}
\alias{atom4R-package}
\title{R package to read/write metadata in Atom XML format.}
\description{
R engine to orchestrate and run geospatial (meta)data workflows
}
\details{
\tabular{ll}{
   Package: \tab atom4R\cr
   Type: \tab Package\cr
   Version: \tab 0.0.9000\cr
   Date: \tab 2020-04-24\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
