% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atom4R_logger.R
\docType{class}
\name{atom4RLogger}
\alias{atom4RLogger}
\title{atom4RLogger}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a simple logger
}
\description{
atom4RLogger
}
\note{
Logger class used internally by atom4R
}
\section{Abstract Methods}{

\describe{
 \item{\code{INFO(text)}}{
   Logger to report information. Used internally
 }
 \item{\code{WARN(text)}}{
   Logger to report warnings. Used internally
 }
 \item{\code{ERROR(text)}}{
   Logger to report errors. Used internally
 }
}
}

\keyword{logger}
