% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomFeed.R
\docType{class}
\name{AtomFeed}
\alias{AtomFeed}
\title{Atom feed class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Atom feed
}
\description{
This class models an atom feed
}
\details{
AtomFeed
}
\section{Fields}{

\describe{
\item{\code{id}}{Identifier}

\item{\code{updated}}{Update date}

\item{\code{title}}{Title}

\item{\code{subtitle}}{Subtitle}

\item{\code{rights}}{Rights (license, use, ...)}

\item{\code{author}}{Author person}

\item{\code{contributor}}{Contributor person}

\item{\code{generator}}{Generator}

\item{\code{icon}}{Icon}

\item{\code{logo}}{Logo}

\item{\code{category}}{Category}

\item{\code{entries}}{List of entries}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to create an Atom
 }
}
}

\examples{
\dontrun{
}

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{atom}
\keyword{feed}
