% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomNamespace.R
\docType{class}
\name{AtomNamespace}
\alias{AtomNamespace}
\title{AtomNamespace}
\format{
\code{ \link[R6]{R6Class}} object.
}
\value{
Object of \code{ \link[R6]{R6Class}} for modelling an Atom Namespace
}
\description{
AtomNamespace

AtomNamespace
}
\note{
ISO class used internally by atom4R for specifying XML namespaces
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ISO}
\keyword{metadata}
\keyword{namespace}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{id}

\item{\code{uri}}{uri}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AtomNamespace-new}{\code{AtomNamespace$new()}}
\item \href{#method-AtomNamespace-getDefinition}{\code{AtomNamespace$getDefinition()}}
\item \href{#method-AtomNamespace-clone}{\code{AtomNamespace$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AtomNamespace-new"></a>}}
\if{latex}{\out{\hypertarget{method-AtomNamespace-new}{}}}
\subsection{Method \code{new()}}{
Initializes an \link{AtomNamespace}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AtomNamespace$new(id, uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{id}

\item{\code{uri}}{uri}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AtomNamespace-getDefinition"></a>}}
\if{latex}{\out{\hypertarget{method-AtomNamespace-getDefinition}{}}}
\subsection{Method \code{getDefinition()}}{
Get definition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AtomNamespace$getDefinition()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a named list defining the namespace
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AtomNamespace-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AtomNamespace-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AtomNamespace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
