% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomPerson.R
\docType{class}
\name{AtomPerson}
\alias{AtomPerson}
\title{Atom Person class}
\format{\code{\link{R6Class}} object.}
\usage{
AtomPerson
}
\value{
Object of \code{\link{R6Class}} for modelling an Atom Person
}
\description{
This class models an Atom Person
}
\details{
AtomPerson
}
\section{Fields}{

\describe{
\item{\code{email}}{}
}}

\note{
Abstract class used internally for person-like classes
}
\section{Methods}{

\describe{
 \item{\code{new(xml, name, uri, email)}}{
   This method is used to create an Atom
 }
 \item{\code{setName(name)}}{
   Set person name
 }
 \item{\code{setUri(uri)}}{
   Set person uri
 }
 \item{\code{setEmail(email)}}{
   Set person email
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Atom}
\keyword{Person}
