% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomAbstractObject.R
\name{cacheAtomClasses}
\alias{cacheAtomClasses}
\title{cacheAtomClasses}
\usage{
cacheAtomClasses()
}
\description{
\code{\link{cacheAtomClasses}} allows to cache the list of
\pkg{atom4R} classes or extended. This is especially required to fasten
the decoding of metadata elements from an XML file. It is called internally
by \pkg{atom4R} the first function \code{\link{getAtomClasses}} is called
and each time the function \code{\link{readAtom}} function is called to
integrate eventually new classes added by user to extend \pkg{atom4R} model
(case of Atom profiles).
}
\examples{
  cacheAtomClasses()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
