% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCMIVocabulary.R
\docType{class}
\name{DCMIVocabulary}
\alias{DCMIVocabulary}
\title{DCMI Vocabulary class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Dublin Core element
}
\description{
This class models an DCMI Vocabulary
}
\details{
DCMIVocabulary
}
\section{Fields}{

\describe{
\item{\code{value}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(doc, format)}}{
   This method is used to read a DCMI vocabulary RDF doc. The format corresponds to
   the RDF format as used by \pkg{rdflib} \code{rdf_parse} function.
 }
 \item{\code{get()}}{
   Runs a Sparql query over the RDF vocabulary to return the vocabulary content.
   Returns an object of class \code{data.frame}
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Core}
\keyword{Dublin}
\keyword{element}
