% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AtomAbstractObject.R
\name{getAtomClasses}
\alias{getAtomClasses}
\title{getAtomClasses}
\usage{
getAtomClasses()
}
\description{
get the list of cached Atom classes
}
\examples{
  getAtomClasses()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
