% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwordClient.R, R/SwordServiceDocument.R
\docType{class}
\name{SwordClient}
\alias{SwordClient}
\alias{SwordServiceDocument}
\title{SwordClient class}
\format{
\code{\link{R6Class}} object.

\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Sword client

Object of \code{\link{R6Class}} for modelling an Sword client
}
\description{
This class models an Sword service client

This class models an Sword service client
}
\details{
SwordClient

SwordClient
}
\section{Methods}{

\describe{
 \item{\code{new(url, user, password, token)}}{
   This method is to instantiate an Sword Client
 }
}


\describe{
 \item{\code{new(url, user, password, token)}}{
   This method is to instantiate an Sword Client
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Atom}
\keyword{Person}
