% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwordClient.R
\docType{data}
\name{SwordDataverseClient}
\alias{SwordDataverseClient}
\title{SwordDataverseClient}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SwordDataverseClient
}
\description{
SwordDataverseClient
}
\keyword{datasets}
