% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwordServiceDocument.R
\docType{class}
\name{SwordServiceDocument}
\alias{SwordServiceDocument}
\title{SwordServiceDocument class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Sword service document
}
\description{
This class models an Sword service document
}
\details{
SwordServiceDocument
}
\note{
class used internally by \pkg{atom4R}
}
\section{Methods}{

\describe{
 \item{\code{new(xml, logger)}}{
   This method is to instantiate an SwordServiceDocument
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Atom}
\keyword{Person}
