% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwordServiceDocument.R
\docType{class}
\name{SwordServiceDocument}
\alias{SwordServiceDocument}
\title{SwordServiceDocument class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling an Sword service document
}
\description{
This class models an Sword service document
}
\details{
SwordServiceDocument
}
\note{
class used internally by \pkg{atom4R}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{Atom}
\keyword{Person}
\section{Super class}{
\code{\link[atom4R:atom4RLogger]{atom4R::atom4RLogger}} -> \code{SwordServiceDocument}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{title}

\item{\code{collections}}{collections}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SwordServiceDocument-new}{\code{SwordServiceDocument$new()}}
\item \href{#method-SwordServiceDocument-getTitle}{\code{SwordServiceDocument$getTitle()}}
\item \href{#method-SwordServiceDocument-getCollections}{\code{SwordServiceDocument$getCollections()}}
\item \href{#method-SwordServiceDocument-clone}{\code{SwordServiceDocument$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="atom4R" data-topic="atom4RLogger" data-id="ERROR"><a href='../../atom4R/html/atom4RLogger.html#method-atom4RLogger-ERROR'><code>atom4R::atom4RLogger$ERROR()</code></a></span></li>
<li><span class="pkg-link" data-pkg="atom4R" data-topic="atom4RLogger" data-id="INFO"><a href='../../atom4R/html/atom4RLogger.html#method-atom4RLogger-INFO'><code>atom4R::atom4RLogger$INFO()</code></a></span></li>
<li><span class="pkg-link" data-pkg="atom4R" data-topic="atom4RLogger" data-id="WARN"><a href='../../atom4R/html/atom4RLogger.html#method-atom4RLogger-WARN'><code>atom4R::atom4RLogger$WARN()</code></a></span></li>
<li><span class="pkg-link" data-pkg="atom4R" data-topic="atom4RLogger" data-id="getClass"><a href='../../atom4R/html/atom4RLogger.html#method-atom4RLogger-getClass'><code>atom4R::atom4RLogger$getClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="atom4R" data-topic="atom4RLogger" data-id="getClassName"><a href='../../atom4R/html/atom4RLogger.html#method-atom4RLogger-getClassName'><code>atom4R::atom4RLogger$getClassName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="atom4R" data-topic="atom4RLogger" data-id="logger"><a href='../../atom4R/html/atom4RLogger.html#method-atom4RLogger-logger'><code>atom4R::atom4RLogger$logger()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SwordServiceDocument-new"></a>}}
\if{latex}{\out{\hypertarget{method-SwordServiceDocument-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \link{SwordServiceDocument} from XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwordServiceDocument$new(xml, logger = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{XMLInternalNode-class} from \pkg{XML}}

\item{\code{logger}}{logger}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SwordServiceDocument-getTitle"></a>}}
\if{latex}{\out{\hypertarget{method-SwordServiceDocument-getTitle}{}}}
\subsection{Method \code{getTitle()}}{
Get title
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwordServiceDocument$getTitle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SwordServiceDocument-getCollections"></a>}}
\if{latex}{\out{\hypertarget{method-SwordServiceDocument-getCollections}{}}}
\subsection{Method \code{getCollections()}}{
Get collections
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwordServiceDocument$getCollections()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SwordServiceDocument-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SwordServiceDocument-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SwordServiceDocument$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
